refactor: add a base class LLMStreamService
